Added chemistry subcategories, Account menu, and Organic Chemistry Menu
We need the HOME button to be a menu closer. The account menu needs to have a place to put in username and password along with a checkbox to say if you're a peer reviewer. The fake documents under Organic chemistry need vote and donate buttons. I wasn't able to get the menu opener to work for other menus so that is a priority.